--- src/main/java/com/dazzle/shop/address/AddreessService.java
@@ -1,5 +1,0 @@
-package com.dazzle.shop.address;
-
-public class AddreessService {
-
-}
