--- src/main/java/com/dazzle/shop/address/impl/AddressDAO.java
@@ -1,5 +1,0 @@
-package com.dazzle.shop.address.impl;
-
-public class AddressDAO {
-
-}
