--- src/main/java/com/dazzle/shop/address/impl/AddressServiceImpl.java
@@ -1,5 +1,0 @@
-package com.dazzle.shop.address.impl;
-
-public class AddressServiceImpl {
-
-}
