--- src/main/java/com/dazzle/shop/admin/AdminService.java
@@ -1,5 +1,0 @@
-package com.dazzle.shop.admin;
-
-public class AdminService {
-
-}
