--- src/main/java/com/dazzle/shop/admin/AdminVO.java
@@ -1,5 +1,0 @@
-package com.dazzle.shop.admin;
-
-public class AdminVO {
-
-}
