--- src/main/java/com/dazzle/shop/admin/impl/AdminDAO.java
@@ -1,5 +1,0 @@
-package com.dazzle.shop.admin.impl;
-
-public class AdminDAO {
-
-}
