--- src/main/java/com/dazzle/shop/board/BoardService.java
@@ -1,5 +1,0 @@
-package com.dazzle.shop.board;
-
-public class BoardService {
-
-}
