--- src/main/java/com/dazzle/shop/board/BoardVO.java
@@ -1,5 +1,0 @@
-package com.dazzle.shop.board;
-
-public class BoardVO {
-
-}
