--- src/main/java/com/dazzle/shop/board/impl/BoardServiceImpl.java
@@ -1,5 +1,0 @@
-package com.dazzle.shop.board.impl;
-
-public class BoardServiceImpl {
-
-}
