+++ src/main/java/com/dazzle/shop/model/address/AddreessService.java
@@ -1,0 +1,5 @@
+package com.dazzle.shop.model.address;
+
+public class AddreessService {
+
+}
