+++ src/main/java/com/dazzle/shop/model/address/impl/AddressDAO.java
@@ -1,0 +1,5 @@
+package com.dazzle.shop.model.address.impl;
+
+public class AddressDAO {
+
+}
