+++ src/main/java/com/dazzle/shop/model/address/impl/AddressServiceImpl.java
@@ -1,0 +1,5 @@
+package com.dazzle.shop.model.address.impl;
+
+public class AddressServiceImpl {
+
+}
