+++ src/main/java/com/dazzle/shop/model/admin/AdminService.java
@@ -1,0 +1,5 @@
+package com.dazzle.shop.model.admin;
+
+public class AdminService {
+
+}
