+++ src/main/java/com/dazzle/shop/model/admin/impl/AdminServiceImpl.java
@@ -1,0 +1,5 @@
+package com.dazzle.shop.model.admin.impl;
+
+public class AdminServiceImpl {
+
+}
