+++ src/main/java/com/dazzle/shop/model/board/BoardService.java
@@ -1,0 +1,5 @@
+package com.dazzle.shop.model.board;
+
+public class BoardService {
+
+}
