+++ src/main/java/com/dazzle/shop/model/board/impl/BoardServiceImpl.java
@@ -1,0 +1,5 @@
+package com.dazzle.shop.model.board.impl;
+
+public class BoardServiceImpl {
+
+}
