+++ src/main/java/com/dazzle/shop/model/order/OrderVO.java
@@ -1,0 +1,5 @@
+package com.dazzle.shop.model.order;
+
+public class OrderVO {
+
+}
