+++ src/main/java/com/dazzle/shop/model/order/impl/OrderDAO.java
@@ -1,0 +1,5 @@
+package com.dazzle.shop.model.order.impl;
+
+public class OrderDAO {
+
+}
