+++ src/main/java/com/dazzle/shop/model/product/ProductVO.java
@@ -1,0 +1,5 @@
+package com.dazzle.shop.model.product;
+
+public class ProductVO {
+
+}
