+++ src/main/java/com/dazzle/shop/model/product/impl/ProductDAO.java
@@ -1,0 +1,5 @@
+package com.dazzle.shop.model.product.impl;
+
+public class ProductDAO {
+
+}
