+++ src/main/java/com/dazzle/shop/model/product/impl/ProductServiceImpl.java
@@ -1,0 +1,5 @@
+package com.dazzle.shop.model.product.impl;
+
+public class ProductServiceImpl {
+
+}
