+++ src/main/java/com/dazzle/shop/model/user/UserService.java
@@ -1,0 +1,5 @@
+package com.dazzle.shop.model.user;
+
+public class UserService {
+
+}
