+++ src/main/java/com/dazzle/shop/model/user/impl/UserDAO.java
@@ -1,0 +1,5 @@
+package com.dazzle.shop.model.user.impl;
+
+public class UserDAO {
+
+}
