--- src/main/java/com/dazzle/shop/order/OrderService.java
@@ -1,5 +1,0 @@
-package com.dazzle.shop.order;
-
-public class OrderService {
-
-}
