--- src/main/java/com/dazzle/shop/order/impl/OrderDAO.java
@@ -1,5 +1,0 @@
-package com.dazzle.shop.order.impl;
-
-public class OrderDAO {
-
-}
