--- src/main/java/com/dazzle/shop/order/impl/OrderServiceImpl.java
@@ -1,5 +1,0 @@
-package com.dazzle.shop.order.impl;
-
-public class OrderServiceImpl {
-
-}
