--- src/main/java/com/dazzle/shop/product/ProductService.java
@@ -1,5 +1,0 @@
-package com.dazzle.shop.product;
-
-public class ProductService {
-
-}
