--- src/main/java/com/dazzle/shop/product/impl/ProductServiceImpl.java
@@ -1,5 +1,0 @@
-package com.dazzle.shop.product.impl;
-
-public class ProductServiceImpl {
-
-}
