--- src/main/java/com/dazzle/shop/user/UserVO.java
@@ -1,5 +1,0 @@
-package com.dazzle.shop.user;
-
-public class UserVO {
-
-}
