--- src/main/java/com/dazzle/shop/user/impl/UserServiceImpl.java
@@ -1,5 +1,0 @@
-package com.dazzle.shop.user.impl;
-
-public class UserServiceImpl {
-
-}
